hold nvidia-fabricmanager version
The cuda drivers are installed via nvidia's ".run" script; whereas, the fabric manager is installed via the deb packages.  This means that if you run `apt-get upgrade`, `apt-get update`, the fabric manager and drivers will not be in sync, and you will get an cuda initialization error.  So, this change tries to fix this issue.