fixed udev rules for nfs-readahead
This was broken as it was split on multiple lines.  This caused errors
for the first three and then applied the read ahead to every device.
The way the file was written also meant that $kernel and $4 were
evaluated by the shell when writing the file as the $'s needed to be
escaped.